requirements.txt
aiofiles==24.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
load-dotenv==0.1.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
playwright==1.44.0
pluggy==1.5.0
pyee==11.1.0
pytest==8.2.2
pytest-base-url==2.1.0
pytest-playwright==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
setuptools==70.1.1
six==1.16.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
